+++ server/server-config/src/main/kotlin/suwayomi/tachidesk/server/ServerConfig.kt
@@ -131,14 +131,14 @@ class ServerConfig(
     val socksProxyHost: MutableStateFlow<String> by StringSetting(
         protoNumber = 5,
         group = SettingGroup.PROXY,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
     )
 
     val socksProxyPort: MutableStateFlow<String> by StringSetting(
         protoNumber = 6,
         group = SettingGroup.PROXY,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
     )
 
@@ -187,7 +187,7 @@ class ServerConfig(
     val electronPath: MutableStateFlow<String> by PathSetting(
         protoNumber = 12,
         group = SettingGroup.WEB_UI,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
         mustExist = true,
         excludeFromBackup = true,
@@ -222,7 +222,7 @@ class ServerConfig(
     val downloadsPath: MutableStateFlow<String> by PathSetting(
         protoNumber = 16,
         group = SettingGroup.DOWNLOADER,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
         mustExist = true,
         excludeFromBackup = true,
@@ -284,6 +284,7 @@ class ServerConfig(
         defaultValue = emptyList(),
         deprecated =
             SettingsRegistry.SettingDeprecated(
+                replaceWith = "extensionStores",
                 message = "Replaced with addExtensionStore and removeExtensionStore mutations",
                 migrateConfigValue = {
                     @Suppress("UNCHECKED_CAST")
@@ -419,6 +420,7 @@ class ServerConfig(
         defaultValue = false,
         deprecated =
             SettingsRegistry.SettingDeprecated(
+                replaceWith = null,
                 message = "Removed - does not do anything",
             ),
     )
@@ -461,7 +463,7 @@ class ServerConfig(
     val backupPath: MutableStateFlow<String> by PathSetting(
         protoNumber = 38,
         group = SettingGroup.BACKUP,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
         mustExist = true,
         excludeFromBackup = true,
@@ -497,7 +499,7 @@ class ServerConfig(
     val localSourcePath: MutableStateFlow<String> by PathSetting(
         protoNumber = 42,
         group = SettingGroup.LOCAL_SOURCE,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
         mustExist = true,
         excludeFromBackup = true,
@@ -514,7 +516,7 @@ class ServerConfig(
     val flareSolverrUrl: MutableStateFlow<String> by StringSetting(
         protoNumber = 44,
         group = SettingGroup.CLOUDFLARE,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "http://localhost:8191",
     )
 
@@ -530,7 +532,7 @@ class ServerConfig(
     val flareSolverrSessionName: MutableStateFlow<String> by StringSetting(
         protoNumber = 46,
         group = SettingGroup.CLOUDFLARE,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "suwayomi",
     )
 
@@ -711,7 +713,7 @@ class ServerConfig(
     val koreaderSyncServerUrl: MutableStateFlow<String> by MigratedConfigValue(
         protoNumber = 59,
         group = SettingGroup.KOREADER_SYNC,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "https://sync.koreader.rocks/",
         deprecated = SettingsRegistry.SettingDeprecated(
             replaceWith = "MOVE TO PREFERENCES",
@@ -765,7 +767,7 @@ class ServerConfig(
     val koreaderSyncDeviceId: MutableStateFlow<String> by MigratedConfigValue(
         protoNumber = 62,
         group = SettingGroup.KOREADER_SYNC,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "",
         deprecated = SettingsRegistry.SettingDeprecated(
             replaceWith = "MOVE TO PREFERENCES",
@@ -909,7 +911,7 @@ class ServerConfig(
     val databaseUrl: MutableStateFlow<String> by StringSetting(
         protoNumber = 70,
         group = SettingGroup.DATABASE,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = "postgresql://localhost:5432/suwayomi",
         excludeFromBackup = true,
     )
@@ -1054,7 +1056,7 @@ class ServerConfig(
         protoNumber = 88,
         defaultValue = "",
         group = SettingGroup.SYNCYOMI,
-        privacySafe = true,
+        privacySafe = false,
     )
 
     val syncYomiApiKey: MutableStateFlow<String> by StringSetting(
@@ -1117,7 +1119,7 @@ class ServerConfig(
     val extensionStores: MutableStateFlow<List<String>> by ListSetting<String>(
         protoNumber = 97,
         group = SettingGroup.EXTENSION,
-        privacySafe = true,
+        privacySafe = false,
         defaultValue = emptyList(),
         requiresRestart = true,
         itemValidator = { url ->