+++ server/src/main/kotlin/suwayomi/tachidesk/global/impl/KcefWebView.kt
@@ -19,6 +19,7 @@ import org.cef.handler.CefLoadHandlerAdapter
 import org.cef.handler.CefRenderHandlerAdapter
 import org.cef.handler.CefRequestHandlerAdapter
 import org.cef.handler.CefResourceRequestHandler
+import org.cef.handler.CefResourceRequestHandlerAdapter
 import org.cef.input.CefTouchEvent
 import org.cef.misc.BoolRef
 import org.cef.network.CefCookie
@@ -27,6 +28,7 @@ import org.cef.network.CefRequest
 import uy.kohesive.injekt.injectLazy
 import xyz.nulldev.androidcompat.webkit.CefHelper
 import xyz.nulldev.androidcompat.webkit.dispose
+import xyz.nulldev.androidcompat.webkit.disposeWithJsHandler
 import xyz.nulldev.androidcompat.webkit.evaluateJavaScript
 import java.awt.Component
 import java.awt.HeadlessException
@@ -192,6 +194,19 @@ class KcefWebView {
         }
     }
 
+    private inner class ResourceRequestHandler : CefResourceRequestHandlerAdapter() {
+        override fun onBeforeResourceLoad(
+            browser: CefBrowser?,
+            frame: CefFrame?,
+            request: CefRequest,
+        ): Boolean {
+            val ua = System.getProperty("http.agent")
+            request.setHeaderByName("user-agent", ua, true)
+            logger.trace { "Using user-agent $ua" }
+            return false
+        }
+    }
+
     private inner class RequestHandler : CefRequestHandlerAdapter() {
         override fun getResourceRequestHandler(
             browser: CefBrowser,
@@ -203,7 +218,7 @@ class KcefWebView {
             disableDefaultHandling: BoolRef,
         ): CefResourceRequestHandler? {
             logger.trace { "Load resource: ${frame.name} - ${request.url}" }
-            return null
+            return ResourceRequestHandler()
         }
     }
 
@@ -282,7 +297,7 @@ class KcefWebView {
         browser?.close(true)
         browser?.dispose()
         browser = null
-        kcefClient?.dispose()
+        kcefClient?.disposeWithJsHandler()
         kcefClient = null
     }