+++ server/src/main/kotlin/suwayomi/tachidesk/manga/impl/download/Downloader.kt
@@ -7,6 +7,7 @@ package suwayomi.tachidesk.manga.impl.download
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+import eu.kanade.tachiyomi.source.local.LocalSource
 import io.github.oshai.kotlinlogging.KLogger
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.CancellationException
@@ -28,6 +29,7 @@ import suwayomi.tachidesk.manga.impl.download.model.DownloadState.Error
 import suwayomi.tachidesk.manga.impl.download.model.DownloadState.Finished
 import suwayomi.tachidesk.manga.impl.download.model.DownloadState.Queued
 import suwayomi.tachidesk.manga.impl.download.model.DownloadUpdate
+import suwayomi.tachidesk.manga.impl.download.model.DownloadUpdateType
 import suwayomi.tachidesk.manga.impl.download.model.DownloadUpdateType.ERROR
 import suwayomi.tachidesk.manga.impl.download.model.DownloadUpdateType.FINISHED
 import suwayomi.tachidesk.manga.impl.download.model.DownloadUpdateType.PAUSED
@@ -60,12 +62,32 @@ class Downloader(
 
     class EmptyChapterException : Exception("Chapter does not have any pages to download")
 
+    private fun notify(
+        immediate: Boolean,
+        update: DownloadUpdate? = null,
+    ) {
+        val isDownloadCanceled = !downloadQueue.contains(update?.downloadQueueItem)
+        if (isDownloadCanceled) {
+            return
+        }
+
+        notifier(immediate, update)
+    }
+
+    private fun notify(
+        immediate: Boolean,
+        type: DownloadUpdateType,
+        download: DownloadQueueItem?,
+    ) {
+        notify(immediate, download?.let { DownloadUpdate(type, download, downloadQueue.indexOf(download)) })
+    }
+
     private suspend fun step(
-        downloadUpdate: DownloadUpdate?,
+        type: DownloadUpdateType,
+        download: DownloadQueueItem?,
         immediate: Boolean,
     ) {
-        val download = downloadUpdate?.downloadQueueItem
-        notifier(immediate, downloadUpdate)
+        notify(immediate, type, download)
         currentCoroutineContext().ensureActive()
         if (download != null && download != availableSourceDownloads.firstOrNull { it.state != Error }) {
             if (download in downloadQueue) {
@@ -94,7 +116,7 @@ class Downloader(
                         }
                     }
             logger.debug { "started" }
-            notifier(false, null)
+            notify(false)
         }
     }
 
@@ -107,7 +129,7 @@ class Downloader(
         logger: KLogger,
         download: DownloadQueueItem,
     ) {
-        notifier(true, DownloadUpdate(FINISHED, download))
+        notify(true, FINISHED, download)
         downloadQueue -= download
         onDownloadFinished()
         logger.debug { "finished" }
@@ -123,6 +145,12 @@ class Downloader(
             val logContext = "${logger.name} - downloadChapter($download))"
             val downloadLogger = KotlinLogging.logger(logContext)
 
+            // Skip local source downloads
+            if (sourceId == LocalSource.ID) {
+                finishDownload(downloadLogger, download)
+                continue
+            }
+
             downloadLogger.debug { "start" }
 
             // handle cases were the downloader was stopped before the finished download could be removed from the queue
@@ -135,7 +163,7 @@ class Downloader(
 
             try {
                 download.state = Downloading
-                step(DownloadUpdate(PROGRESS, download), true)
+                step(PROGRESS, download, true)
 
                 val chapter = getChapterDownloadReadyById(download.chapterId)
 
@@ -146,7 +174,7 @@ class Downloader(
                 download.pageCount = chapter.pageCount
 
                 ChapterDownloadHelper.download(download.mangaId, download.chapterId, download, scope) { downloadChapter, immediate ->
-                    step(downloadChapter?.let { DownloadUpdate(PROGRESS, downloadChapter) }, immediate)
+                    step(PROGRESS, downloadChapter, immediate)
                 }
                 download.state = Finished
                 transaction {
@@ -158,25 +186,27 @@ class Downloader(
                 }
                 finishDownload(downloadLogger, download)
             } catch (e: CancellationException) {
-                logger.debug { "Downloader was stopped" }
+                downloadLogger.debug { "Downloader was stopped" }
                 availableSourceDownloads.filter { it.state == Downloading }.forEach { it.state = Queued }
-                notifier(false, DownloadUpdate(STOPPED, download))
+                notify(false, STOPPED, download)
+            } catch (e: StopDownloadException) {
+                downloadLogger.debug { "Download was dequeued" }
             } catch (e: PauseDownloadException) {
                 downloadLogger.debug { "paused" }
                 download.state = Queued
-                notifier(false, DownloadUpdate(PAUSED, download))
+                notify(false, PAUSED, download)
             } catch (e: EmptyChapterException) {
                 downloadLogger.warn(e) { "failed due to" }
                 download.tries = MAX_RETRIES
                 download.state = Error
-                notifier(false, DownloadUpdate(ERROR, download))
+                notify(false, ERROR, download)
             } catch (e: Exception) {
                 downloadLogger.warn(e) { "failed due to" }
                 download.tries++
                 download.state = Queued
                 if (download.tries >= MAX_RETRIES) {
                     download.state = Error
-                    notifier(false, DownloadUpdate(ERROR, download))
+                    notify(false, ERROR, download)
                 }
             }
         }